Everything in the 'data-fetcher' module is coupled to Axios, so lets rename it to 'axios-fetcher' and use 'data-fetcher' for the underlying generic interface
