+++ packages/sitecore-jss/src/dictionary-service.ts
@@ -2,7 +2,7 @@ import mcache from 'memory-cache';
 import { AxiosDataFetcher } from './axios-fetcher';
 import { fetchData } from './data-api';
 import { DictionaryPhrases, DictionaryServiceData } from './data-models';
-import { HttpJsonFetcher } from './http-request-client';
+import { HttpJsonFetcher } from './data-fetcher';
 
 export interface DictionaryService {
   /**
