+++ packages/sitecore-jss/src/layout-service.ts
@@ -1,8 +1,8 @@
 import { AxiosDataFetcher, AxiosDataFetcherConfig } from './axios-fetcher';
 import { LayoutServiceData, PlaceholderData } from './data-models';
 import { fetchPlaceholderData, fetchRouteData, LayoutServiceConfig } from './data-api';
-import { HttpJsonFetcher } from './http-request-client';
 import { GraphQLRequestClient } from './graphql-request-client';
+import { HttpJsonFetcher } from './data-fetcher';
 import { AxiosRequestConfig, AxiosResponse } from 'axios';
 import { IncomingMessage, ServerResponse } from 'http';
